requirements.txt.txt
streamlit==1.32.2
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
sentence-transformers==2.6.1